+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Hotfix
 
 ### Enhancements
+* KLS-54 - Add subscribe to FTA updates form
 
 ## [2.16.2](https://github.com/uktrade/great-cms/releases/tag/2.16.0)
 ### Enhancements